Move dynamicbeat main loop to dynamicbeat package
