+++ dynamicbeat/pkg/dynamicbeat/run.go
@@ -1,4 +1,4 @@
-package run
+package dynamicbeat
 
 import (
 	"os"
